Refactor DTO renderers
